refactor: separate business logic from info and favorites buttons
