+++ weather-app/src/slices/favoritesSlice.ts
@@ -9,7 +9,7 @@ export const favoritesSlice = createSlice({
   initialState,
   reducers: {
     addFavorite: (state, action: PayloadAction<string>) => {
-      if (!state.includes(action.payload)) state.push(action.payload);
+      if (!state.includes(action.payload)) state.unshift(action.payload);
     },
     removeOneFavorite: (state, action: PayloadAction<string>) => {
       return state.filter((location) => location !== action.payload);
